move claims struct
