+++ crates/defguard_common/src/auth/mod.rs
@@ -1,0 +1,1 @@
+pub mod claims;
