+++ crates/defguard_core/src/grpc/auth.rs
@@ -1,15 +1,13 @@
 use std::sync::{Arc, Mutex};
 
+use defguard_common::auth::claims::{Claims, ClaimsType};
 use defguard_proto::auth::{AuthenticateRequest, AuthenticateResponse, auth_service_server};
 use jsonwebtoken::errors::Error as JWTError;
 use sqlx::PgPool;
 use tonic::{Request, Response, Status};
 
 use crate::{
-    auth::{
-        Claims, ClaimsType,
-        failed_login::{FailedLoginMap, check_failed_logins, log_failed_login_attempt},
-    },
+    auth::failed_login::{FailedLoginMap, check_failed_logins, log_failed_login_attempt},
     db::User,
     server_config,
 };
